// ts/chat-react-core/gen/graphql.ts
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type ChannelAvatarItemPurchase = {
  __typename?: 'ChannelAvatarItemPurchase';
  item: ItemItem;
  itemId: Scalars['ID']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};
// ...
export type ChannelBundlePurchase = {
  __typename?: 'ChannelBundlePurchase';
  bundleName: Scalars['String']['output'];
  streamerCardIds: Array<Scalars['String']['output']>;
  streamerCards?: Maybe<Array<GameLogicStreamerCard>>;
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};
// ...
export type ChannelChannelEvent = {
  __typename?: 'ChannelChannelEvent';
  channelId: Scalars['ID']['output'];
  content: ChannelChannelEventContent;
  createdAt: Scalars['Timestamp']['output'];
  id: Scalars['ID']['output'];
};

export type ChannelChannelEventContent = {
  __typename?: 'ChannelChannelEventContent';
  content?: Maybe<ChannelChannelEventContentContentUnion>;
};
// ...
export type ChannelGiftSubscriptionPurchase = {
  __typename?: 'ChannelGiftSubscriptionPurchase';
  recipientUserIds: Array<Scalars['String']['output']>;
  recipients?: Maybe<Array<ProfileProfile>>;
  tier: Scalars['Int']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};
// ...
export type ChannelStreamerCardPurchase = {
  __typename?: 'ChannelStreamerCardPurchase';
  streamerCard: GameLogicStreamerCard;
  streamerCardId: Scalars['ID']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};

export type ChannelStreamingFeatureStatus = {
  __typename?: 'ChannelStreamingFeatureStatus';
  enabled: Scalars['Boolean']['output'];
  suspension?: Maybe<ChannelSuspension>;
};

export type ChannelSubscriptionPurchase = {
  __typename?: 'ChannelSubscriptionPurchase';
  tier: Scalars['Int']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};
// ...
export type GameLogicStreamerCard = {
  __typename?: 'GameLogicStreamerCard';
  baseCard: GameLogicCard;
  channel: ChannelChannel;
  channelId: Scalars['ID']['output'];
  draft: Scalars['Boolean']['output'];
  facecam: Scalars['String']['output'];
  facecamUrl: Scalars['String']['output'];
  familyId: Scalars['ID']['output'];
  gameId: Scalars['ID']['output'];
  id: Scalars['ID']['output'];
  image: Scalars['String']['output'];
  imageUrl: Scalars['String']['output'];
  name: Scalars['String']['output'];
  saleConfig?: Maybe<StoreV2StreamerCardSaleConfig>;
  video: Scalars['String']['output'];
  videoUrl: Scalars['String']['output'];
};
// ...
export type ItemItem = {
  __typename?: 'ItemItem';
  attributes: AttributeAttributeMap;
  bootstraps?: Maybe<Array<ItemItemBootstrap>>;
  channel?: Maybe<ChannelChannel>;
  channelId: Scalars['ID']['output'];
  children?: Maybe<Array<ItemItem>>;
  consumable: Scalars['Boolean']['output'];
  details?: Maybe<ItemItemDetailsUnion>;
  disabled: Scalars['Boolean']['output'];
  game?: Maybe<GameGame>;
  gameId: Scalars['ID']['output'];
  id: Scalars['ID']['output'];
  inventoryItem: InventoryInventoryItem;
  name: Scalars['String']['output'];
  /** if the parent item is owned, this item is owned as well */
  parentItemId: Scalars['ID']['output'];
  season?: Maybe<GameSeason>;
  seasonId: Scalars['ID']['output'];
  type: ItemItemType;
  /** unlock_item_id what if any item is required to unlock this item for bundles and other consumption */
  unlockItemId: Scalars['ID']['output'];
  updatedAt: Scalars['Timestamp']['output'];
};
// ...
export enum ProfileColor {
  Color_6Ec9F7 = 'COLOR_6EC9F7',
  Color_56F6C0 = 'COLOR_56F6C0',
  Color_63F655 = 'COLOR_63F655',
  Color_8686F9 = 'COLOR_8686F9',
  ColorB26Afb = 'COLOR_B26AFB',
  ColorC0F656 = 'COLOR_C0F656',
  ColorF6Ce56 = 'COLOR_F6CE56',
  ColorF6F656 = 'COLOR_F6F656',
  ColorF76Ef7 = 'COLOR_F76EF7',
  ColorF69856 = 'COLOR_F69856',
  ColorUnspecified = 'COLOR_UNSPECIFIED'
}
// ...
export type ProfileProfile = {
  __typename?: 'ProfileProfile';
  account?: Maybe<AuthAccount>;
  avatarConfig?: Maybe<ProfileProfileAvatarConfig>;
  /**
   * contains a concatenated list of different avatar urls (use avatars object instead)
   * @deprecated field is deprecated
   */
  avatarUrl: Scalars['String']['output'];
  avatars?: Maybe<ProfileProfileAvatars>;
  badges: Array<BadgeBadge>;
  bio: Scalars['String']['output'];
  canChangeUsernameAt?: Maybe<Scalars['Timestamp']['output']>;
  channel?: Maybe<ChannelChannel>;
  discordUsername?: Maybe<Scalars['String']['output']>;
  /** @deprecated field is deprecated */
  displayName: Scalars['String']['output'];
  friends: Array<ProfileProfile>;
  friendshipStatus: FriendsFriendshipStatus;
  isNewUsername: Scalars['Boolean']['output'];
  lastSeen?: Maybe<Scalars['Timestamp']['output']>;
  onlineStatus: ProfilePresenceStatus;
  playedGames: Array<ProfilePlayedGame>;
  preferredColor: ProfileColor;
  settings?: Maybe<ProfileProfileSettings>;
  state: ApiEntityState;
  stats: PlayerStatsPlayerStats;
  temporary: Scalars['Boolean']['output'];
  userId: Scalars['ID']['output'];
  userTag: Scalars['String']['output'];
  usernameHistory?: Maybe<Array<ProfileUsernameChange>>;
  visibility: ProfileProfileVisibility;
};
// ...
export type Subscription = {
  __typename?: 'Subscription';
  /** StreamerService.ChannelActivityEvents */
  channelActivityEventsSubscribe?: Maybe<StreamerChannelActivityEvent>;
  /** ChannelService.ChannelEventStream */
  channelEventsSubscribe?: Maybe<ChannelChannelEvent>;
  /** ChannelService.FollowerCountUpdates */
  channelFollowerCountSubscribe?: Maybe<ChannelFollowerCountEvent>;
  /** ChannelService.LiveStatusUpdates - Subscription APIs */
  channelLiveStatusSubscribe?: Maybe<ChannelLiveStatusEvent>;
  /** ChannelService.ChannelStreamDetailUpdates */
  channelStreamDetailSubscribe?: Maybe<ChannelChannelStreamDetailEvent>;
  /** ChannelService.ViewerCountUpdates */
  channelViewerCountSubscribe?: Maybe<ChannelViewerCountEvent>;
  /** ChatModerationService.StreamAutoModQueue */
  chatAutoModQueueSubscribe?: Maybe<ChatAutoModQueueEvent>;
  /** ChatService.ChatMessageStream */
  chatMessageSubscribe?: Maybe<ChatChatEvent>;
  /** ChannelConfigService.IngestStatsUpdates */
  ingestStatsSubscribe?: Maybe<ChannelIngestStatsEvent>;
  /** MLControllerService.StreamHUDScaleUpdates */
  mlHUDScaleUpdatesSubscribe?: Maybe<MlControllerGetHudScaleResponse>;
  /** ChannelModerationService.ModerationEventsStream */
  moderationEventsSubscribe?: Maybe<ChannelModerationEvent>;
  /** NotificationService.Notifications */
  notificationSubscribe?: Maybe<NotificationNotification>;
  /** PartyService.StreamPartyUpdates */
  partyUpdateSubscribe?: Maybe<PartyPartyUpdateEvent>;
  /** ReactionService.StreamReactions */
  reactionsSubscription?: Maybe<ReactionReactionEvent>;
  /** ChallengeSessionService.ChallengeSessionEvents */
  streamChallengeSessionEvents?: Maybe<ChallengesessionChallengeSessionEvent>;
  /** MatchServiceV2.ChallengeUpdates */
  streamChallengeUpdatesSubscribe?: Maybe<MatchChallengeUpdate>;
  /** StreamerService.StreamDiagnosticsUpdates */
  streamDiagnosticsSubscribe?: Maybe<StreamerStreamDiagnosticsUpdate>;
  /** StreamerService.StreamEvents */
  streamEventsSubscribe?: Maybe<StreamerStreamEvent>;
  /** MatchServiceV2.StreamSpectatorCoordinationEvents */
  streamSpectatorCoordinationEventsSubscribe?: Maybe<MatchStreamSpectatorCoordinationEvent>;
  /** ChannelConfigService.StreamStatusUpdates */
  streamStatusSubscribe?: Maybe<ChannelStreamStatusEvent>;
  /** MatchServiceV2.TopCardUpdates */
  streamTopActiveCardsSubscribe?: Maybe<MatchTopCardsUpdate>;
  /** FeatureFlagService.StreamConfigUpdates */
  subscribeFeatureFlagConfigUpdates?: Maybe<FlagFeatureFlagConfig>;
};
// ...
export type ChatChannelEventsSubscriptionSubscriptionVariables = Exact<{
  channelId: Scalars['ID']['input'];
}>;


export type ChatChannelEventsSubscriptionSubscription = { __typename?: 'Subscription', channelEventsSubscribe?: { __typename?: 'ChannelChannelEvent', channelId: string, createdAt: string, id: string, content: { __typename?: 'ChannelChannelEventContent', content?: { __typename?: 'ChannelAvatarItemPurchase', itemId: string, item: { __typename?: 'ItemItem', id: string, name: string }, user?: { __typename?: 'ProfileProfile', userId: string, preferredColor: ProfileColor, userTag: string } | null } | { __typename?: 'ChannelBundlePurchase', bundleName: string, userId: string, streamerCards?: Array<{ __typename?: 'GameLogicStreamerCard', id: string, channelId: string, baseCard: { __typename?: 'GameLogicCard', id: string, name: string } }> | null, user?: { __typename?: 'ProfileProfile', userId: string, preferredColor: ProfileColor, userTag: string } | null } | { __typename?: 'ChannelGiftSubscriptionPurchase', userId: string, user?: { __typename?: 'ProfileProfile', userId: string, preferredColor: ProfileColor, userTag: string, badges: Array<{ __typename?: 'BadgeBadge', type: BadgeBadgeType, level: number }> } | null, recipients?: Array<{ __typename?: 'ProfileProfile', userId: string, preferredColor: ProfileColor, userTag: string, badges: Array<{ __typename?: 'BadgeBadge', type: BadgeBadgeType, level: number }> }> | null } | { __typename?: 'ChannelStreamerCardPurchase', userId: string, streamerCard: { __typename?: 'GameLogicStreamerCard', id: string, channelId: string, baseCard: { __typename?: 'GameLogicCard', id: string, name: string } }, user?: { __typename?: 'ProfileProfile', userId: string, preferredColor: ProfileColor, userTag: string } | null } | { __typename?: 'ChannelSubscriptionPurchase', userId: string, tier: number, user?: { __typename?: 'ProfileProfile', userId: string, preferredColor: ProfileColor, userTag: string, badges: Array<{ __typename?: 'BadgeBadge', type: BadgeBadgeType, level: number }> } | null } | { __typename?: 'ChannelSubscriptionRenewal' } | null } } | null };
// ...
export const ChannelEventContentProfileFragmentDoc = gql`
    fragment ChannelEventContentProfile on ProfileProfile {
  preferredColor
  userId
  userTag
}
    `;
export const ChannelEventSubscriptionContentProfileFragmentDoc = gql`
    fragment ChannelEventSubscriptionContentProfile on ProfileProfile {
  badges(channel_id: $channelId) {
    type
    level
  }
}
    `;
export const ChannelEventAvatarItemPurchaseItemFragmentDoc = gql`
    fragment ChannelEventAvatarItemPurchaseItem on ItemItem {
  id
  name
}
    `;
// ...
export const ChatChannelEventsSubscriptionDocument = gql`
    subscription ChatChannelEventsSubscription($channelId: ID!) {
  channelEventsSubscribe(channelId: $channelId) {
    channelId
    createdAt
    id
    content {
      content {
        ... on ChannelAvatarItemPurchase {
          itemId
          item {
            ...ChannelEventAvatarItemPurchaseItem
          }
          user {
            userId
            ...ChannelEventContentProfile
          }
        }
        ... on ChannelSubscriptionPurchase {
          userId
          tier
          user {
            userId
            ...ChannelEventContentProfile
            ...ChannelEventSubscriptionContentProfile
          }
        }
        ... on ChannelBundlePurchase {
          bundleName
          userId
          streamerCards {
            id
            channelId
            baseCard {
              id
              name
            }
          }
          user {
            userId
            ...ChannelEventContentProfile
          }
        }
        ... on ChannelStreamerCardPurchase {
          streamerCard {
            id
            channelId
            baseCard {
              id
              name
            }
          }
          userId
          user {
            userId
            ...ChannelEventContentProfile
          }
        }
        ... on ChannelGiftSubscriptionPurchase {
          userId
          user {
            userId
            ...ChannelEventContentProfile
            ...ChannelEventSubscriptionContentProfile
          }
          recipients {
            userId
            ...ChannelEventContentProfile
            ...ChannelEventSubscriptionContentProfile
          }
        }
      }
    }
  }
}
    ${ChannelEventAvatarItemPurchaseItemFragmentDoc}
${ChannelEventContentProfileFragmentDoc}
${ChannelEventSubscriptionContentProfileFragmentDoc}`;